--- MVVMSidekick/MVVMSidekick.Shared/Behaviors/BehaviorHelper.cs
@@ -1,15 +1,0 @@
-﻿#if NETFX_CORE
-
-
-#else
-using System.Windows.Controls;
-using System.Windows.Data;
-
-#endif
-
-namespace MVVMSidekick.Behaviors
-{
-
-
-
-}
