+++ MVVMSidekick/MVVMSidekick.Shared/Behaviors/ListenToEventRouterTrigger.cs
@@ -24,164 +24,164 @@ using MVVMSidekick.Common;
 namespace MVVMSidekick.Behaviors
 {
 
-#if NETFX_CORE
-	public class ListenToEventRouterTriggerBehavior : BehaviorBase
+//#if NETFX_CORE
+//	public class ListenToEventRouterTriggerBehavior : BehaviorBase
 
-	{
-
-
-		public ListenToEventRouterTriggerBehavior()
-		{
-			var binding = new Binding();
-			binding.Path = new PropertyPath(nameof(ObjectTypeFilterAssemblyQualifiedName));
-			binding.Mode = BindingMode.TwoWay;
-			binding.Converter = TypeNameStringToTypeConverter.Instance;
-			binding.TargetNullValue = typeof(object);
-			binding.UpdateSourceTrigger = UpdateSourceTrigger.PropertyChanged;
-			BindingOperations.SetBinding(this, EventObjectTypeProperty, binding);
-		}
+//	{
 
 
-
-		public String ObjectTypeFilterAssemblyQualifiedName
-		{
-			get { return (String)GetValue(ObjectTypeFilterAssemblyQualifiedNameProperty); }
-			set { SetValue(ObjectTypeFilterAssemblyQualifiedNameProperty, value); }
-		}
-
-		// Using a DependencyProperty as the backing store for ObjectTypeFilterAssemblyQualifiedName.  This enables animation, styling, binding, etc...
-		public static readonly DependencyProperty ObjectTypeFilterAssemblyQualifiedNameProperty =
-			DependencyProperty.Register("ObjectTypeFilterAssemblyQualifiedName", typeof(String), typeof(ListenToEventRouterTriggerBehavior), new PropertyMetadata(typeof(object).AssemblyQualifiedName));
-
-
-		private object InvokeActions(object e)
-		{
-			foreach (var item in Actions)
-			{
-				try
-				{
-					item.Execute(this, e);
-				}
-				catch (Exception)
-				{
-					//throw;
-				}
-			}
-			return e;
-		}
-
-		public ObservableCollection<Microsoft.Xaml.Interactivity.IAction> Actions
-		{
-			get { return (ObservableCollection<Microsoft.Xaml.Interactivity.IAction>)GetValue(ActionsProperty); }
-			set { SetValue(ActionsProperty, value); }
-		}
-
-		// Using a DependencyProperty as the backing store for Actions.  This enables animation, styling, binding, etc...
-		public static readonly DependencyProperty ActionsProperty =
-			DependencyProperty.Register("Actions", typeof(ObservableCollection<Microsoft.Xaml.Interactivity.IAction>), typeof(ListenToEventRouterTriggerBehavior), new PropertyMetadata(new ObservableCollection<IAction>()));
+//		public ListenToEventRouterTriggerBehavior()
+//		{
+//			var binding = new Binding();
+//			binding.Path = new PropertyPath(nameof(ObjectTypeFilterAssemblyQualifiedName));
+//			binding.Mode = BindingMode.TwoWay;
+//			binding.Converter = TypeNameStringToTypeConverter.Instance;
+//			binding.TargetNullValue = typeof(object);
+//			binding.UpdateSourceTrigger = UpdateSourceTrigger.PropertyChanged;
+//			BindingOperations.SetBinding(this, EventObjectTypeProperty, binding);
+//		}
 
 
+
+//		public String ObjectTypeFilterAssemblyQualifiedName
+//		{
+//			get { return (String)GetValue(ObjectTypeFilterAssemblyQualifiedNameProperty); }
+//			set { SetValue(ObjectTypeFilterAssemblyQualifiedNameProperty, value); }
+//		}
+
+//		// Using a DependencyProperty as the backing store for ObjectTypeFilterAssemblyQualifiedName.  This enables animation, styling, binding, etc...
+//		public static readonly DependencyProperty ObjectTypeFilterAssemblyQualifiedNameProperty =
+//			DependencyProperty.Register("ObjectTypeFilterAssemblyQualifiedName", typeof(String), typeof(ListenToEventRouterTriggerBehavior), new PropertyMetadata(typeof(object).AssemblyQualifiedName));
+
+
+//		private object InvokeActions(object e)
+//		{
+//			foreach (var item in Actions)
+//			{
+//				try
+//				{
+//					item.Execute(this, e);
+//				}
+//				catch (Exception)
+//				{
+//					//throw;
+//				}
+//			}
+//			return e;
+//		}
+
+//		public ObservableCollection<Microsoft.Xaml.Interactivity.IAction> Actions
+//		{
+//			get { return (ObservableCollection<Microsoft.Xaml.Interactivity.IAction>)GetValue(ActionsProperty); }
+//			set { SetValue(ActionsProperty, value); }
+//		}
+
+//		// Using a DependencyProperty as the backing store for Actions.  This enables animation, styling, binding, etc...
+//		public static readonly DependencyProperty ActionsProperty =
+//			DependencyProperty.Register("Actions", typeof(ObservableCollection<Microsoft.Xaml.Interactivity.IAction>), typeof(ListenToEventRouterTriggerBehavior), new PropertyMetadata(new ObservableCollection<IAction>()));
 
 
 
 
 
-		public override void Attach(DependencyObject associatedObject)
-		{
-			ExchangeTheSubscribedRouter(this, EventRouter);
-
-			base.Attach(associatedObject);
-		}
 
 
+//		public override void Attach(DependencyObject associatedObject)
+//		{
+//			ExchangeTheSubscribedRouter(this, EventRouter);
 
-		public override void Detach()
-		{
-			base.Detach();
-			if (_oldSubscrption != null)
-			{
-				_oldSubscrption.Dispose();
-				_oldSubscrption = null;
-			}
-		}
-
-
-		public EventRouter EventRouter
-		{
-			get { return (EventRouter)GetValue(EventRouterProperty); }
-			set { SetValue(EventRouterProperty, value); }
-		}
-
-		// Using a DependencyProperty as the backing store for EventRouter.  This enables animation, styling, binding, etc...
-		public static readonly DependencyProperty EventRouterProperty =
-			DependencyProperty.Register("EventRouter", typeof(EventRouter), typeof(ListenToEventRouterTriggerBehavior), new PropertyMetadata(null,
-				(o, e) =>
-				{
-					var t = o as ListenToEventRouterTriggerBehavior;
-					ExchangeTheSubscribedRouter(t, e.NewValue as EventRouter);
-				}));
+//			base.Attach(associatedObject);
+//		}
 
 
 
-
-		/// <summary>
-		/// Gets or sets the type of the event object.
-		/// </summary>
-		/// <value>
-		/// The type of the event object.
-		/// </value>
-		public System.Type EventObjectType
-		{
-			get { return (Type)GetValue(EventObjectTypeProperty); }
-			set { SetValue(EventObjectTypeProperty, value); }
-		}
-
-		// Using a DependencyProperty as the backing store for EventObjectType.  This enables animation, styling, binding, etc...
-		public static readonly DependencyProperty EventObjectTypeProperty =
-			DependencyProperty.Register("EventObjectType", typeof(Type), typeof(ListenToEventRouterTriggerBehavior), new PropertyMetadata(typeof(object)));
+//		public override void Detach()
+//		{
+//			base.Detach();
+//			if (_oldSubscrption != null)
+//			{
+//				_oldSubscrption.Dispose();
+//				_oldSubscrption = null;
+//			}
+//		}
 
 
+//		public EventRouter EventRouter
+//		{
+//			get { return (EventRouter)GetValue(EventRouterProperty); }
+//			set { SetValue(EventRouterProperty, value); }
+//		}
 
-		/// <summary>
-		/// Gets or sets the name of the event routing.
-		/// </summary>
-		/// <value>
-		/// The name of the event routing.
-		/// </value>
-		public string EventRoutingName
-		{
-			get { return (string)GetValue(EventRoutingNameProperty); }
-			set { SetValue(EventRoutingNameProperty, value); }
-		}
-
-
-		// Using a DependencyProperty as the backing store for EventRoutingName.  This enables animation, styling, binding, etc...
-		public static readonly DependencyProperty EventRoutingNameProperty =
-			DependencyProperty.Register("EventRoutingName", typeof(string), typeof(ListenToEventRouterTriggerBehavior), new PropertyMetadata(null));
+//		// Using a DependencyProperty as the backing store for EventRouter.  This enables animation, styling, binding, etc...
+//		public static readonly DependencyProperty EventRouterProperty =
+//			DependencyProperty.Register("EventRouter", typeof(EventRouter), typeof(ListenToEventRouterTriggerBehavior), new PropertyMetadata(null,
+//				(o, e) =>
+//				{
+//					var t = o as ListenToEventRouterTriggerBehavior;
+//					ExchangeTheSubscribedRouter(t, e.NewValue as EventRouter);
+//				}));
 
 
 
 
-		IDisposable _oldSubscrption;
+//		/// <summary>
+//		/// Gets or sets the type of the event object.
+//		/// </summary>
+//		/// <value>
+//		/// The type of the event object.
+//		/// </value>
+//		public System.Type EventObjectType
+//		{
+//			get { return (Type)GetValue(EventObjectTypeProperty); }
+//			set { SetValue(EventObjectTypeProperty, value); }
+//		}
 
-		static void ExchangeTheSubscribedRouter(ListenToEventRouterTriggerBehavior trigger, EventRouter newRouter)
-		{
-
-			var targetEventRouter = newRouter ?? EventRouter.Instance;
-			var query = targetEventRouter.GetEventChannel<object>()
-				.Where(x => string.IsNullOrEmpty(trigger.EventRoutingName) || trigger.EventRoutingName == x.EventName);
+//		// Using a DependencyProperty as the backing store for EventObjectType.  This enables animation, styling, binding, etc...
+//		public static readonly DependencyProperty EventObjectTypeProperty =
+//			DependencyProperty.Register("EventObjectType", typeof(Type), typeof(ListenToEventRouterTriggerBehavior), new PropertyMetadata(typeof(object)));
 
 
-			var old = Interlocked.Exchange(
-				ref trigger._oldSubscrption,
-				query.Subscribe(e => trigger.InvokeActions(e)).MakeFinalizableDisposable());
 
-			old.Dispose();
+//		/// <summary>
+//		/// Gets or sets the name of the event routing.
+//		/// </summary>
+//		/// <value>
+//		/// The name of the event routing.
+//		/// </value>
+//		public string EventRoutingName
+//		{
+//			get { return (string)GetValue(EventRoutingNameProperty); }
+//			set { SetValue(EventRoutingNameProperty, value); }
+//		}
 
-		}
 
-	}
-#endif
+//		// Using a DependencyProperty as the backing store for EventRoutingName.  This enables animation, styling, binding, etc...
+//		public static readonly DependencyProperty EventRoutingNameProperty =
+//			DependencyProperty.Register("EventRoutingName", typeof(string), typeof(ListenToEventRouterTriggerBehavior), new PropertyMetadata(null));
+
+
+
+
+//		IDisposable _oldSubscrption;
+
+//		static void ExchangeTheSubscribedRouter(ListenToEventRouterTriggerBehavior trigger, EventRouter newRouter)
+//		{
+
+//			var targetEventRouter = newRouter ?? EventRouter.Instance;
+//			var query = targetEventRouter.GetEventChannel<object>()
+//				.Where(x => string.IsNullOrEmpty(trigger.EventRoutingName) || trigger.EventRoutingName == x.EventName);
+
+
+//			var old = Interlocked.Exchange(
+//				ref trigger._oldSubscrption,
+//				query.Subscribe(e => trigger.InvokeActions(e)).MakeFinalizableDisposable());
+
+//			old.Dispose();
+
+//		}
+
+//	}
+//#endif
 
 #if NETFX_CORE
 
@@ -207,41 +207,12 @@ namespace MVVMSidekick.Behaviors
 		}
 	}
 #endif
-#if NETFX_CORE
-	public class ListenToEventRouterDataBehavior : BehaviorBase
-	{
-		public ListenToEventRouterDataBehavior()
-		{
-			var binding = new Binding();
-			binding.Path = new PropertyPath(nameof(ObjectTypeFilterAssemblyQualifiedName));
-			binding.Mode = BindingMode.TwoWay;
-			binding.Converter = TypeNameStringToTypeConverter.Instance;
-			binding.TargetNullValue = typeof(object);
-			binding.UpdateSourceTrigger = UpdateSourceTrigger.PropertyChanged;
-			BindingOperations.SetBinding(this, EventObjectTypeProperty, binding);
-		}
 
-
-
-		public String ObjectTypeFilterAssemblyQualifiedName
-		{
-			get { return (String)GetValue(ObjectTypeFilterAssemblyQualifiedNameProperty); }
-			set { SetValue(ObjectTypeFilterAssemblyQualifiedNameProperty, value); }
-		}
-
-		// Using a DependencyProperty as the backing store for ObjectTypeFilterAssemblyQualifiedName.  This enables animation, styling, binding, etc...
-		public static readonly DependencyProperty ObjectTypeFilterAssemblyQualifiedNameProperty =
-			DependencyProperty.Register("ObjectTypeFilterAssemblyQualifiedName", typeof(String), typeof(ListenToEventRouterDataBehavior), new PropertyMetadata(
-				typeof(object).AssemblyQualifiedName));
-
-
-
-#else
 
     /// <summary>
     /// ListenToEventRouterData Behavior
     /// </summary>
-    public sealed class ListenToEventRouterDataBehavior : Behavior<FrameworkElement>, IDisposable
+    public sealed class ListenToEventRouterDataBehavior : Behavior<FrameworkElement>,IDisposable
 	{
 
 
@@ -280,8 +251,6 @@ namespace MVVMSidekick.Behaviors
 			base.OnDetaching();
 		}
 
-
-#endif
 
 
         /// <summary>
@@ -363,7 +332,7 @@ namespace MVVMSidekick.Behaviors
 		}
 
 
-#if !NETFX_CORE
+
         /// <summary>
         /// Releases unmanaged and - optionally - managed resources.
         /// </summary>
@@ -372,7 +341,7 @@ namespace MVVMSidekick.Behaviors
 			_oldSubscrption?.Dispose();
 		}
 
-#endif
+
 
 
 		/// <summary>
@@ -395,7 +364,7 @@ namespace MVVMSidekick.Behaviors
         /// The name of the event routing.
         /// </value>
         public static readonly DependencyProperty EventRoutingNameProperty =
-			DependencyProperty.Register("EventRoutingName", typeof(string), typeof(ListenToEventRouterDataBehavior), new PropertyMetadata(null));
+			DependencyProperty.Register(nameof(EventRoutingName), typeof(string), typeof(ListenToEventRouterDataBehavior), new PropertyMetadata(null));
