+++ MVVMSidekick/MVVMSidekick.Shared/Commands/ICommandWithViewModel.cs
@@ -1,0 +1,69 @@
+﻿// ***********************************************************************
+// Assembly         : MVVMSidekick_Wp8
+// Author           : waywa
+// Created          : 05-17-2014
+//
+// Last Modified By : waywa
+// Last Modified On : 01-04-2015
+// ***********************************************************************
+// <copyright file="Commands.cs" company="">
+//     Copyright ©  2012
+// </copyright>
+// <summary></summary>
+// ***********************************************************************
+using System.Windows.Input;
+using MVVMSidekick.ViewModels;
+#if NETFX_CORE
+
+#elif WPF
+using System.Windows;
+using System.Windows.Controls;
+using System.Windows.Data;
+using System.Collections.Concurrent;
+using System.Windows.Navigation;
+
+using System.Windows.Controls.Primitives;
+
+#elif SILVERLIGHT_5 || SILVERLIGHT_4
+using System.Windows;
+using System.Windows.Controls;
+using System.Windows.Data;
+using System.Windows.Navigation;
+using System.Windows.Controls.Primitives;
+using Microsoft.Runtime.CompilerServices;
+
+#elif WINDOWS_PHONE_8 || WINDOWS_PHONE_7
+using System.Windows;
+using System.Windows.Controls;
+using Microsoft.Phone.Controls;
+using System.Windows.Data;
+using System.Windows.Navigation;
+using System.Windows.Controls.Primitives;
+#endif
+
+
+
+
+
+
+namespace MVVMSidekick
+{
+
+    namespace Commands
+    {
+        /// <summary>
+        /// 带有VM的Command接口
+        /// </summary>
+        public interface ICommandWithViewModel : ICommand
+		{
+			/// <summary>
+			/// Gets or sets the view model.
+			/// </summary>
+			/// <value>The view model.</value>
+			BindableBase ViewModel { get; set; }
+		}
+
+
+	}
+
+}
