+++ MVVMSidekick/MVVMSidekick.Shared/Common/CommonHelper.cs
@@ -1,0 +1,13 @@
+﻿using System;
+
+namespace MVVMSidekick.Common
+{
+    public static class CommonHelper
+	{
+		public static FinalizableDisposable<TInnerDisposable> MakeFinalizableDisposable<TInnerDisposable>(this TInnerDisposable innerDisposable) where TInnerDisposable : class, IDisposable
+		{
+			return new FinalizableDisposable<TInnerDisposable>(innerDisposable);
+		}
+
+	}
+}
