+++ MVVMSidekick/MVVMSidekick.Shared/Common/InstanceCounter.cs
@@ -1,0 +1,15 @@
+﻿using System.Threading;
+
+namespace MVVMSidekick.Common
+{
+    public class InstanceCounter
+	{
+		int count;
+
+		public int Click()
+		{
+			return Interlocked.Increment(ref count);
+		}
+
+	}
+}
