+++ MVVMSidekick/MVVMSidekick.Shared/Services/ITypeSpecifiedServiceLocator.cs
@@ -1,0 +1,128 @@
+﻿// ***********************************************************************
+// Assembly         : MVVMSidekick_Wp8
+// Author           : waywa
+// Created          : 05-17-2014
+//
+// Last Modified By : waywa
+// Last Modified On : 01-04-2015
+// ***********************************************************************
+// <copyright file="Services.cs" company="">
+//     Copyright ©  2012
+// </copyright>
+// <summary></summary>
+// ***********************************************************************
+using System;
+using System.Threading.Tasks;
+#if NETFX_CORE
+
+
+
+#elif WPF
+
+
+
+#elif SILVERLIGHT_5 || SILVERLIGHT_4
+using System.Windows;
+using System.Windows.Controls;
+using System.Windows.Data;
+using System.Windows.Navigation;
+using System.Windows.Controls.Primitives;
+#elif WINDOWS_PHONE_8 || WINDOWS_PHONE_7
+using System.Windows;
+using System.Windows.Controls;
+using Microsoft.Phone.Controls;
+using System.Windows.Data;
+using System.Windows.Navigation;
+using System.Windows.Controls.Primitives;
+
+#endif
+
+
+namespace MVVMSidekick
+{
+
+
+    namespace Services
+    {
+        /// <summary>
+        /// Interface ITypeSpecifiedServiceLocator
+        /// </summary>
+        /// <typeparam name="TService">The type of the t service.</typeparam>
+        public interface ITypeSpecifiedServiceLocator<TService>
+        {
+            /// <summary>
+            /// Determines whether the specified name has instance.
+            /// </summary>
+            /// <param name="name">The name.</param>
+            /// <returns><c>true</c> if the specified name has instance; otherwise, <c>false</c>.</returns>
+            bool HasInstance(string name = "");
+            /// <summary>
+            /// Determines whether the specified name is asynchronous.
+            /// </summary>
+            /// <param name="name">The name.</param>
+            /// <returns><c>true</c> if the specified name is asynchronous; otherwise, <c>false</c>.</returns>
+            bool IsAsync(string name = "");
+            /// <summary>
+            /// Registers the specified instance.
+            /// </summary>
+            /// <param name="instance">The instance.</param>
+            /// <returns>ServiceLocatorEntryStruct&lt;TService&gt;.</returns>
+            ServiceLocatorEntryStruct<TService> Register(TService instance);
+            /// <summary>
+            /// Registers the specified name.
+            /// </summary>
+            /// <param name="name">The name.</param>
+            /// <param name="instance">The instance.</param>
+            /// <returns>ServiceLocatorEntryStruct&lt;TService&gt;.</returns>
+            ServiceLocatorEntryStruct<TService> Register(string name, TService instance);
+            /// <summary>
+            /// Registers the specified factory.
+            /// </summary>
+            /// <param name="factory">The factory.</param>
+            /// <param name="isAlwaysCreatingNew">if set to <c>true</c> [always new].</param>
+            /// <returns>ServiceLocatorEntryStruct&lt;TService&gt;.</returns>
+            ServiceLocatorEntryStruct<TService> Register(Func<object, TService> factory, bool isAlwaysCreatingNew = true);
+            /// <summary>
+            /// Registers the specified name.
+            /// </summary>
+            /// <param name="name">The name.</param>
+            /// <param name="factory">The factory.</param>
+            /// <param name="isAlwaysCreatingNew">if set to <c>true</c> [always new].</param>
+            /// <returns>ServiceLocatorEntryStruct&lt;TService&gt;.</returns>
+            ServiceLocatorEntryStruct<TService> Register(string name, Func<object, TService> factory, bool isAlwaysCreatingNew = true);
+            /// <summary>
+            /// Resolves the specified name.
+            /// </summary>
+            /// <param name="name">The name.</param>
+            /// <param name="parameter">The parameter.</param>
+            /// <returns>TService.</returns>
+            TService Resolve(string name = null, object parameter = null);
+            /// <summary>
+            /// Registers the specified asynchronous factory.
+            /// </summary>
+            /// <param name="asyncFactory">The asynchronous factory.</param>
+            /// <param name="isAlwaysCreatingNew">if set to <c>true</c> [always new].</param>
+            /// <returns>ServiceLocatorEntryStruct&lt;TService&gt;.</returns>
+            ServiceLocatorEntryStruct<TService> Register(Func<object, Task<TService>> asyncFactory, bool isAlwaysCreatingNew = true);
+            /// <summary>
+            /// Registers the specified name.
+            /// </summary>
+            /// <param name="name">The name.</param>
+            /// <param name="asyncFactory">The asynchronous factory.</param>
+            /// <param name="isAlwaysCreatingNew">if set to <c>true</c> [always new].</param>
+            /// <returns>ServiceLocatorEntryStruct&lt;TService&gt;.</returns>
+            ServiceLocatorEntryStruct<TService> Register(string name, Func<object, Task<TService>> asyncFactory, bool isAlwaysCreatingNew = true);
+            /// <summary>
+            /// Resolves the asynchronous.
+            /// </summary>
+            /// <param name="name">The name.</param>
+            /// <param name="parameter">The parameter.</param>
+            /// <returns>Task&lt;TService&gt;.</returns>
+            Task<TService> ResolveAsync(string name = null, object parameter = null);
+        }
+
+
+
+
+    }
+}
