+++ MVVMSidekick/MVVMSidekick.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
-VisualStudioVersion = 15.0.26430.6
+VisualStudioVersion = 15.0.26430.14
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CommonCode", "CommonCode\CommonCode.csproj", "{4D755FF7-84F5-4D29-A771-55F520F87814}"
 EndProject
@@ -37,8 +37,6 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MVVMSidekick_UAP10", "MVVMSidekick_UAP10\MVVMSidekick_UAP10.csproj", "{E3F7BD12-E681-4552-8F3B-473F325B1A2B}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MVVMSidekick_net45", "MVVMSidekick_net45\MVVMSidekick_net45.csproj", "{0DFFBBC1-9A4D-4183-80EB-37DA3F507D9E}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MVVMSidekick_UAP10.Test", "MVVMSidekick_UAP10.Test\MVVMSidekick_UAP10.Test.csproj", "{D1B1DCFB-6348-4CCE-84A2-7A0A2ACD0908}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Validation", "Validation", "{CFCE545E-BDCE-4442-B76F-507F75389931}"
 EndProject
@@ -172,32 +170,6 @@ Global
 		{0DFFBBC1-9A4D-4183-80EB-37DA3F507D9E}.Release|x64.Build.0 = Release|Any CPU
 		{0DFFBBC1-9A4D-4183-80EB-37DA3F507D9E}.Release|x86.ActiveCfg = Release|Any CPU
 		{0DFFBBC1-9A4D-4183-80EB-37DA3F507D9E}.Release|x86.Build.0 = Release|Any CPU
-		{D1B1DCFB-6348-4CCE-84A2-7A0A2ACD0908}.Debug|Any CPU.ActiveCfg = Debug|x86
-		{D1B1DCFB-6348-4CCE-84A2-7A0A2ACD0908}.Debug|ARM.ActiveCfg = Debug|ARM
-		{D1B1DCFB-6348-4CCE-84A2-7A0A2ACD0908}.Debug|ARM.Build.0 = Debug|ARM
-		{D1B1DCFB-6348-4CCE-84A2-7A0A2ACD0908}.Debug|ARM.Deploy.0 = Debug|ARM
-		{D1B1DCFB-6348-4CCE-84A2-7A0A2ACD0908}.Debug|Mixed Platforms.ActiveCfg = Debug|x86
-		{D1B1DCFB-6348-4CCE-84A2-7A0A2ACD0908}.Debug|Mixed Platforms.Build.0 = Debug|x86
-		{D1B1DCFB-6348-4CCE-84A2-7A0A2ACD0908}.Debug|Mixed Platforms.Deploy.0 = Debug|x86
-		{D1B1DCFB-6348-4CCE-84A2-7A0A2ACD0908}.Debug|x64.ActiveCfg = Debug|x64
-		{D1B1DCFB-6348-4CCE-84A2-7A0A2ACD0908}.Debug|x64.Build.0 = Debug|x64
-		{D1B1DCFB-6348-4CCE-84A2-7A0A2ACD0908}.Debug|x64.Deploy.0 = Debug|x64
-		{D1B1DCFB-6348-4CCE-84A2-7A0A2ACD0908}.Debug|x86.ActiveCfg = Debug|x86
-		{D1B1DCFB-6348-4CCE-84A2-7A0A2ACD0908}.Debug|x86.Build.0 = Debug|x86
-		{D1B1DCFB-6348-4CCE-84A2-7A0A2ACD0908}.Debug|x86.Deploy.0 = Debug|x86
-		{D1B1DCFB-6348-4CCE-84A2-7A0A2ACD0908}.Release|Any CPU.ActiveCfg = Release|x86
-		{D1B1DCFB-6348-4CCE-84A2-7A0A2ACD0908}.Release|ARM.ActiveCfg = Release|ARM
-		{D1B1DCFB-6348-4CCE-84A2-7A0A2ACD0908}.Release|ARM.Build.0 = Release|ARM
-		{D1B1DCFB-6348-4CCE-84A2-7A0A2ACD0908}.Release|ARM.Deploy.0 = Release|ARM
-		{D1B1DCFB-6348-4CCE-84A2-7A0A2ACD0908}.Release|Mixed Platforms.ActiveCfg = Release|x86
-		{D1B1DCFB-6348-4CCE-84A2-7A0A2ACD0908}.Release|Mixed Platforms.Build.0 = Release|x86
-		{D1B1DCFB-6348-4CCE-84A2-7A0A2ACD0908}.Release|Mixed Platforms.Deploy.0 = Release|x86
-		{D1B1DCFB-6348-4CCE-84A2-7A0A2ACD0908}.Release|x64.ActiveCfg = Release|x64
-		{D1B1DCFB-6348-4CCE-84A2-7A0A2ACD0908}.Release|x64.Build.0 = Release|x64
-		{D1B1DCFB-6348-4CCE-84A2-7A0A2ACD0908}.Release|x64.Deploy.0 = Release|x64
-		{D1B1DCFB-6348-4CCE-84A2-7A0A2ACD0908}.Release|x86.ActiveCfg = Release|x86
-		{D1B1DCFB-6348-4CCE-84A2-7A0A2ACD0908}.Release|x86.Build.0 = Release|x86
-		{D1B1DCFB-6348-4CCE-84A2-7A0A2ACD0908}.Release|x86.Deploy.0 = Release|x86
 		{17375DDC-753B-42E8-B14F-6C3DFE9C5A34}.Debug|Any CPU.ActiveCfg = Debug|x86
 		{17375DDC-753B-42E8-B14F-6C3DFE9C5A34}.Debug|ARM.ActiveCfg = Debug|ARM
 		{17375DDC-753B-42E8-B14F-6C3DFE9C5A34}.Debug|ARM.Build.0 = Debug|ARM
@@ -304,7 +276,6 @@ Global
 		{4E76A573-00FF-4DDB-88EB-B6B32A1D950B} = {D631B7B0-E263-4EF0-8749-9EB9B1AFE0C8}
 		{E3F7BD12-E681-4552-8F3B-473F325B1A2B} = {C2833223-E4B1-418D-9668-D8DC53213FCA}
 		{0DFFBBC1-9A4D-4183-80EB-37DA3F507D9E} = {C2833223-E4B1-418D-9668-D8DC53213FCA}
-		{D1B1DCFB-6348-4CCE-84A2-7A0A2ACD0908} = {C5921E6D-635D-4D9B-A1C2-C84661D66CD4}
 		{CFCE545E-BDCE-4442-B76F-507F75389931} = {9A63239D-4816-4C82-B408-AAB6B0D5FE39}
 		{17375DDC-753B-42E8-B14F-6C3DFE9C5A34} = {CFCE545E-BDCE-4442-B76F-507F75389931}
 		{CE6F8C18-0894-4F12-A9BF-FD82A1D210F9} = {C2833223-E4B1-418D-9668-D8DC53213FCA}
